refactor(config): move `PeppySchema` to a dedicated `schema` module
