+++ crates/config-internal/src/launcher.rs
@@ -9,5 +9,5 @@ pub use parse::PeppyLauncherParser;
 pub use types::{
     Deployment, DeploymentGitSource, DeploymentInstance, DeploymentLocalSource,
     DeploymentRepoSource, DeploymentSource, DeploymentUrlSource, FrameworkOverrides, Name,
-    PeppyLauncher, PeppySchema,
+    PeppyLauncher,
 };
